backend/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
coverage==6.0.2
distlib==0.3.3
Django==3.2.8
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==9.3.1
filelock==3.3.1
identify==2.3.0
idna==3.3
nodeenv==1.6.0
platformdirs==2.4.0
pre-commit==2.15.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
virtualenv==20.8.1
